GICv3: LPI/ITS support as MSI controller
GICv3 supports the presence of an Interrupt Translation Service
(ITS) that can behave as a MSI controller.

The ITS allows a peripheral to generate a Locality-specific
Peripheral Interrupt (LPI) by writing to a specific doorbell
location. This interrupt is then exposed by the GIC.

This patch adds support for LPIs, and plugs the ITS into the MSI
framework.

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>