ARM/ARM64: arch-timer: fix arch_timer_probed logic
Commit c387f07e6205 (clocksource: arm_arch_timer: Discard unavailable
timers correctly) changed the way the driver makes sure both the memory
and system-register timers have been probed before finalizing the probing.

There is a interesting flaw in this logic that leads to this final step
never to be executed. Things seems to work pretty well until something
actually needs the data that is produced during this final stage.

For example, KVM explodes on the first run of a guest when executed on
a platform that has both memory and sysreg nodes (Juno, for example).

Just fix the damned logic, and enjoy booting VMs again.

Tested on a Juno system.

Cc: Sudeep Holla <sudeep.holla at arm.com>
Cc: Stephen Boyd <sboyd at codeaurora.org>
Cc: Mark Rutland <mark.rutland at arm.com>
Cc: Daniel Lezcano <daniel.lezcano at linaro.org>
Cc: Christoffer Dall <christoffer.dall at linaro.org>
Reported-by: Riku Voipio <riku.voipio at linaro.org>
Signed-off-by: Marc Zyngier <marc.zyngier at arm.com>