Updated Workaround patch for Thunder Errata 22558.
Previous fix was stoping on-going simd instruction before returning from isr,
however on multi-node we have seen x0 register is getting corrupted in isr
functions(which is used by the gic-v3 function to pass the argument).

This is fixed by adding work-around at the beginnig of isr so that ongoing simd
instruction will be stopped and will not corrupt any registers.

Tested on both single-node and multi-node.

Signed-off-by: Ganapatrao Kulkarni <ganapatrao.kulkarni@caviumnetworks.com>