ARM64/KVM: fix ThunderX pass 1.0 bug #23331
As ThunderX pass 1.0 have some problems with LR registers. this patch attempts to fix that by
emulating GICv3 virtual CPU interface in KVM.

Patch description:

1. Notification of virtual interrupt pending will be sent if cup is in VEL1 and ( LR has a pending entry OR hcr_el2.vi is set).
2. Software maintains all LR state no LR write or read will be done to hardware.
3. Setting tall1/0 results in IAR and EOIR to trap.
4. So before jumping to VEL1 if a pending interrupt exists I will set VI to 1.
5. While handling IAR if I see no interrupts pending VI is set to 0. That way I can control the virtual interrupt notification.
6. Rest of the patch details are self explanatory.

NOTE: i am not enclosing this patch with any CONFIGS becuase, i don't know wether we need to upstream it or not,
even if we want to upstream we need to follow dynamic binary changes, in both cases we don't need to have any new CONFIG
for this.

Signed-off-by: Tirumalesh Chalamarla <tchalamarla@cavium.com>

Conflicts:
	arch/arm64/kvm/sys_regs.c