pci, thunderx: Run resource fixup also for ACPI
Commitid:

 e8eb88c143c1 arm64/Thunder pci: Apply Thunder pci quirks only if needed

broke pci initialization for ACPI. &thunder_pcie_ecam_probed was only
set for devicetree and thus not applied for ACPI.

Now:

 c22be811cf10 Fixes to get ACPI based kernel booting. Temporary fix to get going.

just removed the check not considering an ACPI kernel can also be
boot with devicetree which broke e8eb88c143c1.

So the proper fix is to enable &thunder_pcie_ecam_probed also for
ACPI. This is solved by moving atomic_set() to thunder_pcie_config()
which is called in case of devicetree and ACPI. This makes disabling
the check obsolete.

Signed-off-by: Robert Richter <rrichter@cavium.com>