net: thunderx: acpi: Get mac address from acpi table
This patch adds code to set the mac address of the device as provided
by acpi tables. This is similar to the implementation for devicetree
in of_get_mac_address(). The table is searched for the device property
entries "mac-address", "local-mac-address" and "address" in that
order. The address is provided in a u64 variable and must contain a
valid 6 bytes-len mac addr.

Based on a patch from Narinder Dhillon <ndhillon@cavium.com>.

Signed-off-by: Narinder Dhillon <ndhillon@cavium.com>
Signed-off-by: Robert Richter <rrichter@cavium.com>